For #3203: Updates edit bookmarks to have user friendly names Co-authored-by: Colin Lee <mncolinlee@gmail.com> Co-authored-by: Sawyer Blatz <sdblatz@gmail.com>"
